refactor: 💡 refactor referral to old referral references
