+++ lib/modules/ovc_intervention/submodules/ovc_referral/ovc_referral_pages/ovc_house_referral_pages/pages/ovc_household_referral_view.dart
@@ -9,8 +9,8 @@ import 'package:kb_mobile_app/models/events.dart';
 import 'package:kb_mobile_app/models/intervention_card.dart';
 import 'package:kb_mobile_app/models/ovc_household.dart';
 import 'package:kb_mobile_app/modules/ovc_intervention/components/ovc_household_top_header.dart';
-import 'package:kb_mobile_app/core/components/referrals/referral_detailed_card.dart';
-import 'package:kb_mobile_app/core/components/referrals/referral_outcome_card.dart';
+import 'package:kb_mobile_app/core/components/referrals_old/referral_detailed_card.dart';
+import 'package:kb_mobile_app/core/components/referrals_old/referral_outcome_card.dart';
 import 'package:kb_mobile_app/modules/ovc_intervention/submodules/ovc_referral/ovc_referral_pages/ovc_house_referral_pages/constants/ovc_household_referral_constant.dart';
 import 'package:provider/provider.dart';
